requirements.txt
Django==5.0.3
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
black==24.3.0
pip~=23.3.2
attrs~=23.1.0
platformdirs~=4.2.0
colorama~=0.4.6
packaging~=24.0
pathspec~=0.12.1
click~=8.1.7
aiohttp~=3.9.1
multidict~=6.0.4
asgiref~=3.7.2
sqlparse~=0.4.4
requests~=2.31.0
setuptools~=65.5.0
djangorestframework~=3.14.0
uritemplate~=4.1.1
pytz~=2024.1
PyJWT~=2.8.0